+++ src/Tools/Tool_RBF_DistanceField.hpp
@@ -10,12 +10,12 @@ public:
 private:
 	void reset() override;
 
-	void addAnchorPoint(const glm::vec3& pos) override;
+	void addAnchorPoint(const glm::ivec3& pos) override;
 	void evaluateRBFOnWorld(std::function<void(const glm::ivec3 & pos)> whatToDoWithPos) override;
 	void ImGui_Window() override;
 
 	bool condition(float d);
 
 private:
-	std::vector<glm::vec3> m_anchorPts;
+	std::vector<glm::ivec3> m_anchorPts;
 };
