+++ src/Tools/Tool_RBF_HeightMap.hpp
@@ -10,12 +10,12 @@ public:
 private:
 	void reset() override;
 
-	void addAnchorPoint(const glm::vec3& pos) override;
+	void addAnchorPoint(const glm::ivec3& pos) override;
 	void evaluateRBFOnWorld(std::function<void(const glm::ivec3 & pos)> whatToDoWithPos) override;
 	void ImGui_Window() override;
 
-	bool condition(float h, const glm::vec3& pos);
+	bool condition(float h, const glm::ivec3& pos);
 
 private:
-	std::vector<glm::vec2> m_anchorPts;
+	std::vector<glm::ivec2> m_anchorPts;
 };
